refactor(Dashboard): use molecules for charts, add snohomish config
