+++ src/flavors/pbdurham/config.json
@@ -41,7 +41,11 @@
   "dashboard": {
     "title": "How should we spend 2.4 million dollars?",
     "datasetId": "pbdurham",
-    "datasetOwner": "smartercleanup"
+    "datasetOwner": "smartercleanup",
+    "surveyMetrics": {
+      "categories": true,
+      "demographics": true
+    }
   },
   "map": {
     "geolocation_enabled": true,
